+++ packages/expo-screen-capture/android/.gitignore
