+++ packages/expo-screen-capture/android/src/main/AndroidManifest.xml
