+++ packages/expo-screen-capture/babel.config.js
