+++ packages/expo-screen-capture/ios/EXScreenCapture.podspec
