+++ packages/expo-screen-capture/ios/EXScreenCapture/EXScreenCaptureModule.h
