+++ packages/expo-screen-capture/package.json
