+++ packages/expo-screen-capture/src/ExpoScreenCapture.ts
