+++ packages/expo-screen-capture/src/ExpoScreenCapture.web.ts
