+++ packages/expo-screen-capture/src/ScreenCapture.ts
