+++ packages/expo-screen-capture/src/__tests__/ScreenCapture-test.js
