+++ packages/expo-screen-capture/tsconfig.json
